requirements.txt
fastapi
uvicorn
openai
pytest
pydantic
pydantic-settings
python-dotenv
